requirements.txt
asgiref==3.3.2
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.7.0rc2
Django==3.2.7
django-cors-headers==3.7.0
django-dbrouter==0.1.1
django-filter==2.4.0
django-templated-mail==1.1.1
django-utils-six==2.0
djangorestframework==3.12.2
djangorestframework-datatables==0.6.0
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-nested-routers==0.92.5
drf-rw-serializers==1.0.4
drf-writable-nested==0.6.2
drf-yasg==1.20.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
lxml==4.6.3
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.9
psycopg2==2.8.6
pyarmor==6.7.2
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-docx==0.8.10
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
